refactor: separate facade support concerns
