+++ src/anonymizer/interface/anonymizer.py
@@ -1,11 +1,12 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+"""Owns the `Anonymizer` facade: wiring config to engine workflows and returning results."""
+
 from __future__ import annotations
 
 import logging
 import os
-import re
 import time
 import uuid
 from collections import Counter
@@ -24,26 +25,8 @@ from anonymizer.config.anonymizer_config import (
 )
 from anonymizer.config.replace_strategies import Substitute
 from anonymizer.engine.constants import (
-    COL_ANY_HIGH_LEAKED,
-    COL_ATTRIBUTE_FIDELITY_INVALID_ENTITIES,
-    COL_ATTRIBUTE_FIDELITY_VALID,
     COL_DETECTED_ENTITIES,
-    COL_DETECTION_INVALID_ENTITIES,
-    COL_DETECTION_VALID,
-    COL_FINAL_ENTITIES,
-    COL_JUDGE_EVALUATION,
-    COL_LEAKAGE_MASS,
-    COL_NEEDS_HUMAN_REVIEW,
-    COL_RELATIONAL_CONSISTENCY_INVALID_RELATIONS,
-    COL_RELATIONAL_CONSISTENCY_VALID,
-    COL_REPLACED_TEXT,
-    COL_REWRITTEN_TEXT,
-    COL_TAGGED_TEXT,
     COL_TEXT,
-    COL_TYPE_FIDELITY_INVALID_REPLACEMENTS,
-    COL_TYPE_FIDELITY_VALID,
-    COL_UTILITY_SCORE,
-    COL_WEIGHTED_LEAKAGE_RATE,
     DEFAULT_ENTITY_LABELS,
 )
 from anonymizer.engine.detection.detection_workflow import EntityDetectionWorkflow
@@ -52,7 +35,7 @@ from anonymizer.engine.evaluation.replace.attribute_fidelity_judge import Attrib
 from anonymizer.engine.evaluation.replace.relational_consistency_judge import RelationalConsistencyJudgeWorkflow
 from anonymizer.engine.evaluation.replace.type_fidelity_judge import TypeFidelityJudgeWorkflow
 from anonymizer.engine.io.reader import read_input
-from anonymizer.engine.ndd.adapter import FailedRecord, NddAdapter
+from anonymizer.engine.ndd.adapter import NddAdapter
 from anonymizer.engine.ndd.model_loader import (
     load_default_model_providers,
     parse_model_configs,
@@ -64,7 +47,13 @@ from anonymizer.engine.replace.replace_runner import ReplacementWorkflow
 from anonymizer.engine.resolved_input import ResolvedInput
 from anonymizer.engine.rewrite.rewrite_workflow import RewriteWorkflow
 from anonymizer.interface.errors import InvalidConfigError
+from anonymizer.interface.output_columns import (
+    build_user_dataframe,
+    rename_output_columns,
+    unrename_output_columns,
+)
 from anonymizer.interface.results import AnonymizerResult, PreviewResult
+from anonymizer.interface.run_telemetry import build_anonymizer_event
 from anonymizer.logging import LOG_INDENT, configure_logging, reapply_log_levels
 from anonymizer.measurement import (
     record_record_metrics,
@@ -72,21 +61,18 @@ from anonymizer.measurement import (
     stage_timer,
 )
 from anonymizer.telemetry import (
-    NOT_APPLICABLE,
-    AnonymizerEvent,
     TaskEnum,
     TaskStatusEnum,
     TelemetryHandler,
     _telemetry_enabled,
-    avg_tokens_per_record,
-    classify_model_host,
-    collect_model_hosts,
-    sort_join_aliases,
 )
 
 if TYPE_CHECKING:
     import pandas as pd
     from data_designer.config.config_builder import DataDesignerConfigBuilder
+
+__all__ = ["Anonymizer"]
+
 
 logger = logging.getLogger("anonymizer")
 
@@ -396,16 +382,16 @@ class Anonymizer:
             except ValueError as exc:
                 raise InvalidConfigError(str(exc)) from exc
             text_column = output.resolved_text_column
-            internal_df = _unrename_output_columns(output.trace_dataframe, resolved_text_column=text_column)
+            internal_df = unrename_output_columns(output.trace_dataframe, resolved_text_column=text_column)
             rewrite_result = self._rewrite_runner.evaluate(
                 internal_df,
                 model_configs=self._model_configs,
                 selected_models=self._selected_models.evaluate,
                 privacy_goal=rewrite_config,
             )
-            renamed_trace = _rename_output_columns(rewrite_result.dataframe, resolved_text_column=text_column)
+            renamed_trace = rename_output_columns(rewrite_result.dataframe, resolved_text_column=text_column)
             return AnonymizerResult(
-                dataframe=_build_user_dataframe(renamed_trace, resolved_text_column=text_column),
+                dataframe=build_user_dataframe(renamed_trace, resolved_text_column=text_column),
                 trace_dataframe=renamed_trace,
                 resolved_text_column=text_column,
                 failed_records=rewrite_result.failed_records,
@@ -434,16 +420,16 @@ class Anonymizer:
         # '__nemo_anonymizer_text_input__'). The judge prompts reference the
         # internal names, so reverse the rename before the DD call and re-apply
         # it on the result.
-        internal_df = _unrename_output_columns(output.trace_dataframe, resolved_text_column=text_column)
+        internal_df = unrename_output_columns(output.trace_dataframe, resolved_text_column=text_column)
         replace_result = self._replace_runner.evaluate(
             internal_df,
             replace_method=replace_method,
             model_configs=self._model_configs,
             selected_models=self._selected_models.evaluate,
         )
-        renamed_trace = _rename_output_columns(replace_result.dataframe, resolved_text_column=text_column)
+        renamed_trace = rename_output_columns(replace_result.dataframe, resolved_text_column=text_column)
         return AnonymizerResult(
-            dataframe=_build_user_dataframe(renamed_trace, resolved_text_column=text_column),
+            dataframe=build_user_dataframe(renamed_trace, resolved_text_column=text_column),
             trace_dataframe=renamed_trace,
             resolved_text_column=text_column,
             failed_records=replace_result.failed_records,
@@ -622,7 +608,7 @@ class Anonymizer:
             for f in all_failures:
                 logger.debug("  %s (%s: %s)", f.record_id, f.step, f.reason)
         text_col = context.resolved_text_column
-        renamed_trace = _rename_output_columns(final_df, resolved_text_column=text_col)
+        renamed_trace = rename_output_columns(final_df, resolved_text_column=text_col)
         logger.info("🎉 Pipeline complete — %d records processed, %d total failures", num_records, len(all_failures))
         record_record_metrics(
             final_df,
@@ -632,7 +618,7 @@ class Anonymizer:
             validation_max_entities_per_call=config.detect.validation_max_entities_per_call,
         )
         return AnonymizerResult(
-            dataframe=_build_user_dataframe(renamed_trace, resolved_text_column=text_col),
+            dataframe=build_user_dataframe(renamed_trace, resolved_text_column=text_col),
             trace_dataframe=renamed_trace,
             resolved_text_column=text_col,
             failed_records=all_failures,
@@ -673,12 +659,14 @@ class Anonymizer:
         try:
             if not getattr(config, "emit_telemetry", True):
                 return
-            # Short-circuit before _build_telemetry_event so we don't pay the
+            # Short-circuit before build_anonymizer_event so we don't pay the
             # tiktoken cost on every record when telemetry is globally disabled
             # via NEMO_TELEMETRY_ENABLED=false.
             if not _telemetry_enabled():
                 return
-            event = self._build_telemetry_event(
+            event = build_anonymizer_event(
+                selected_models=self._selected_models,
+                resolved_providers=self._resolved_providers,
                 task=task,
                 status=status,
                 config=config,
@@ -696,77 +684,6 @@ class Anonymizer:
                 handler.enqueue(event)
         except Exception:  # noqa: BLE001 - best-effort
             logger.debug("Failed to emit telemetry event", exc_info=True)
-
-    def _build_telemetry_event(
-        self,
-        *,
-        task: TaskEnum,
-        status: TaskStatusEnum,
-        config: AnonymizerConfig,
-        data: AnonymizerInput,
-        input_df: pd.DataFrame,
-        result: AnonymizerResult | None,
-        duration_sec: float,
-    ) -> AnonymizerEvent:
-        """Construct an AnonymizerEvent from the current pipeline state."""
-        total_records = int(len(input_df))
-        failed = list(result.failed_records) if result is not None else []
-        failure_count = len(failed)
-        success_count = max(total_records - failure_count, 0)
-
-        avg_tokens = -1
-        if total_records > 0 and COL_TEXT in input_df.columns:
-            avg_tokens = avg_tokens_per_record(input_df[COL_TEXT].astype(str))
-
-        transformation_type = _transformation_type_string(config)
-        rewrite = config.rewrite
-        substitute = config.replace if isinstance(config.replace, Substitute) else None
-
-        models = _collect_step_models(
-            selected=self._selected_models,
-            has_substitute=substitute is not None,
-            has_rewrite=rewrite is not None,
-        )
-        failure_counts = _collect_failure_counts(failed)
-        hosts = _resolve_model_hosts(self._resolved_providers)
-
-        return AnonymizerEvent(
-            task=task,
-            task_status=status,
-            job_duration_sec=duration_sec,
-            num_input_records=total_records,
-            num_success_records=success_count,
-            num_failure_records=failure_count,
-            avg_tokens_per_record=avg_tokens,
-            transformation_type=transformation_type,
-            custom_data_summary_provided=bool(data.data_summary),
-            custom_privacy_goal_provided=_custom_privacy_goal_provided(rewrite),
-            custom_substitute_instructions_provided=bool(substitute is not None and substitute.instructions),
-            max_repair_iterations=(rewrite.max_repair_iterations if rewrite is not None else -1),
-            strict_entity_protection=(rewrite.strict_entity_protection if rewrite is not None else False),
-            repair_iterations_triggered=_repair_iterations_triggered(failed, rewrite is not None),
-            entity_detector_model=models["entity_detector"],
-            entity_validator_model=models["entity_validator"],
-            entity_augmenter_model=models["entity_augmenter"],
-            latent_detector_model=models["latent_detector"],
-            replacement_generator_model=models["replacement_generator"],
-            domain_classifier_model=models["domain_classifier"],
-            disposition_analyzer_model=models["disposition_analyzer"],
-            meaning_extractor_model=models["meaning_extractor"],
-            qa_generator_model=models["qa_generator"],
-            rewriter_model=models["rewriter"],
-            evaluator_model=models["evaluator"],
-            repairer_model=models["repairer"],
-            model_hosts=hosts,
-            entity_detection_failure_count=failure_counts["entity_detection"],
-            latent_detection_failure_count=failure_counts["latent_detection"],
-            replace_map_generation_failure_count=failure_counts["replace_map_generation"],
-            rewrite_pipeline_failure_count=failure_counts["rewrite_pipeline"],
-            rewrite_evaluate_failure_count=failure_counts["rewrite_evaluate"],
-            rewrite_repair_failure_count=failure_counts["rewrite_repair"],
-            rewrite_final_judge_failure_count=failure_counts["rewrite_final_judge"],
-            unknown_step_failure_count=failure_counts["unknown"],
-        )
 
 
 def _unwrap_entities(raw: object) -> list:
@@ -827,218 +744,3 @@ def _resolve_model_providers(
     if not raw_providers:
         raise ValueError("model_providers must contain at least one provider.")
     return [ModelProvider.model_validate(provider) for provider in raw_providers]
-
-
-def _rename_output_columns(df: pd.DataFrame, *, resolved_text_column: str) -> pd.DataFrame:
-    """Rename internal column names to user-facing names."""
-    rename_map: dict[str, str] = {}
-    if COL_TEXT in df.columns:
-        rename_map[COL_TEXT] = resolved_text_column
-    if COL_REPLACED_TEXT in df.columns:
-        rename_map[COL_REPLACED_TEXT] = f"{resolved_text_column}_replaced"
-    if COL_TAGGED_TEXT in df.columns:
-        rename_map[COL_TAGGED_TEXT] = f"{resolved_text_column}_with_spans"
-    if COL_REWRITTEN_TEXT in df.columns:
-        rename_map[COL_REWRITTEN_TEXT] = f"{resolved_text_column}_rewritten"
-    if not rename_map:
-        return df
-    return df.rename(columns=rename_map)
-
-
-def _unrename_output_columns(df: pd.DataFrame, *, resolved_text_column: str) -> pd.DataFrame:
-    """Reverse of :func:`_rename_output_columns`.
-
-    Converts user-facing column names (``biography``, ``biography_replaced``, …)
-    back to the internal names (``__nemo_anonymizer_text_input__``, …) that the
-    judges' prompt templates reference. No-op if the dataframe is already in
-    internal form (``COL_TEXT`` already present).
-    """
-    if COL_TEXT in df.columns:
-        return df
-    rename_map: dict[str, str] = {}
-    if resolved_text_column in df.columns:
-        rename_map[resolved_text_column] = COL_TEXT
-    if f"{resolved_text_column}_replaced" in df.columns:
-        rename_map[f"{resolved_text_column}_replaced"] = COL_REPLACED_TEXT
-    if f"{resolved_text_column}_with_spans" in df.columns:
-        rename_map[f"{resolved_text_column}_with_spans"] = COL_TAGGED_TEXT
-    if f"{resolved_text_column}_rewritten" in df.columns:
-        rename_map[f"{resolved_text_column}_rewritten"] = COL_REWRITTEN_TEXT
-    if not rename_map:
-        return df
-    return df.rename(columns=rename_map)
-
-
-def _build_user_dataframe(trace_dataframe: pd.DataFrame, *, resolved_text_column: str) -> pd.DataFrame:
-    """Filter trace dataframe to the public column set for the active mode.
-
-    Replace:     {text_col}, {text_col}_replaced, {text_col}_with_spans, final_entities,
-                 optional judge verdict columns when available
-    Rewrite:     {text_col}, {text_col}_rewritten, utility_score, leakage_mass, weighted_leakage_rate,
-                 any_high_leaked, needs_human_review
-    Detect-only: {text_col}, {text_col}_with_spans, final_entities
-    """
-    t = trace_dataframe
-    text_col = resolved_text_column
-
-    if f"{text_col}_rewritten" in t.columns:
-        allowed = {
-            text_col,
-            f"{text_col}_rewritten",
-            COL_UTILITY_SCORE,
-            COL_LEAKAGE_MASS,
-            COL_WEIGHTED_LEAKAGE_RATE,
-            COL_ANY_HIGH_LEAKED,
-            COL_NEEDS_HUMAN_REVIEW,
-            COL_DETECTION_VALID,  # only present after evaluate()
-            COL_DETECTION_INVALID_ENTITIES,  # only present after evaluate()
-            COL_JUDGE_EVALUATION,  # only present after evaluate()
-        }
-    elif f"{text_col}_replaced" in t.columns:
-        allowed = {
-            text_col,
-            f"{text_col}_replaced",
-            f"{text_col}_with_spans",
-            COL_FINAL_ENTITIES,
-            COL_DETECTION_VALID,
-            COL_DETECTION_INVALID_ENTITIES,
-            COL_TYPE_FIDELITY_VALID,
-            COL_TYPE_FIDELITY_INVALID_REPLACEMENTS,
-            COL_RELATIONAL_CONSISTENCY_VALID,
-            COL_RELATIONAL_CONSISTENCY_INVALID_RELATIONS,
-            COL_ATTRIBUTE_FIDELITY_VALID,
-            COL_ATTRIBUTE_FIDELITY_INVALID_ENTITIES,
-        }
-    else:
-        allowed = {
-            text_col,
-            f"{text_col}_with_spans",
-            COL_FINAL_ENTITIES,
-        }
-
-    return t[[col for col in t.columns if col in allowed]].copy()
-
-
-# ----------------------------------------------------------------- telemetry helpers
-
-
-_REWRITE_REPAIR_RE = re.compile(r"^rewrite-repair-(\d+)$")
-_REWRITE_EVALUATE_RE = re.compile(r"^rewrite-evaluate-(\d+)$")
-
-
-def _transformation_type_string(config: AnonymizerConfig) -> str:
-    """Map AnonymizerConfig to the schema's transformationType value.
-
-    Schema accepts exactly one of: ``annotate``, ``redact``, ``hash``,
-    ``substitute``, ``rewrite``. AnonymizerConfig's validator enforces exactly one
-    of replace/rewrite, so one of these branches always fires.
-    """
-    if config.rewrite is not None:
-        return "rewrite"
-    # The four ReplaceMethodBase subclasses (Annotate, Redact, Hash, Substitute)
-    # lowercase directly to their schema values.
-    return type(config.replace).__name__.lower()
-
-
-def _custom_privacy_goal_provided(rewrite: object | None) -> bool:
-    """Detect whether the user supplied a non-default privacy_goal.
-
-    ``Rewrite.populate_default_privacy_goal`` always populates a default if the
-    user passed None, so we treat the default protect/preserve text as "not custom".
-    """
-    if rewrite is None or rewrite.privacy_goal is None:  # type: ignore[union-attr]
-        return False
-    from anonymizer.config.rewrite import DEFAULT_PRESERVE_TEXT, DEFAULT_PROTECT_TEXT
-
-    goal = rewrite.privacy_goal  # type: ignore[union-attr]
-    return goal.protect != DEFAULT_PROTECT_TEXT or goal.preserve != DEFAULT_PRESERVE_TEXT
-
-
-def _collect_step_models(
-    *,
-    selected,  # ModelSelection
-    has_substitute: bool,
-    has_rewrite: bool,
-) -> dict[str, str]:
-    """Project the user's model selection into the schema's step-keyed shape."""
-    det = selected.detection
-    rewrite = selected.rewrite
-    replace = selected.replace
-    return {
-        "entity_detector": det.entity_detector or NOT_APPLICABLE,
-        "entity_validator": sort_join_aliases(det.entity_validator or []),
-        "entity_augmenter": det.entity_augmenter or NOT_APPLICABLE,
-        # latent_detector only runs in rewrite mode
-        "latent_detector": (det.latent_detector or NOT_APPLICABLE) if has_rewrite else NOT_APPLICABLE,
-        # replacement_generator only runs in Substitute mode
-        "replacement_generator": replace.replacement_generator if has_substitute else NOT_APPLICABLE,
-        # All rewrite-only roles
-        "domain_classifier": rewrite.domain_classifier if has_rewrite else NOT_APPLICABLE,
-        "disposition_analyzer": rewrite.disposition_analyzer if has_rewrite else NOT_APPLICABLE,
-        "meaning_extractor": rewrite.meaning_extractor if has_rewrite else NOT_APPLICABLE,
-        "qa_generator": rewrite.qa_generator if has_rewrite else NOT_APPLICABLE,
-        "rewriter": rewrite.rewriter if has_rewrite else NOT_APPLICABLE,
-        "evaluator": rewrite.evaluator if has_rewrite else NOT_APPLICABLE,
-        "repairer": rewrite.repairer if has_rewrite else NOT_APPLICABLE,
-    }
-
-
-def _step_to_field(step: str) -> str:
-    """Map a FailedRecord.step (workflow_name) to a schema failure-count field key."""
-    match step:
-        case "entity-detection":
-            return "entity_detection"
-        case "latent-entity-detection":
-            return "latent_detection"
-        case "replace-map-generation":
-            return "replace_map_generation"
-        case "rewrite-pipeline":
-            return "rewrite_pipeline"
-        case "rewrite-final-judge":
-            return "rewrite_final_judge"
-        case _ if _REWRITE_EVALUATE_RE.match(step):
-            return "rewrite_evaluate"
-        case _ if _REWRITE_REPAIR_RE.match(step):
-            return "rewrite_repair"
-        case _:
-            return "unknown"
-
-
-def _collect_failure_counts(failed: list[FailedRecord]) -> dict[str, int]:
-    """Aggregate FailedRecord.step values into per-workflow failure counts."""
-    counts = {
-        "entity_detection": 0,
-        "latent_detection": 0,
-        "replace_map_generation": 0,
-        "rewrite_pipeline": 0,
-        "rewrite_evaluate": 0,
-        "rewrite_repair": 0,
-        "rewrite_final_judge": 0,
-        "unknown": 0,
-    }
-    for fr in failed:
-        counts[_step_to_field(fr.step)] += 1
-    return counts
-
-
-def _repair_iterations_triggered(failed: list[FailedRecord], is_rewrite: bool) -> int:
-    """Count distinct repair iterations observed in FailedRecord step names.
-
-    Falls back to -1 when the run wasn't a rewrite. Returns 0 when rewrite ran
-    but no failures surfaced from repair iterations — note that this undercounts
-    repair iterations that completed without producing FailedRecord entries. A
-    follow-up could plumb a richer signal up from the rewrite workflow.
-    """
-    if not is_rewrite:
-        return -1
-    iterations: set[int] = set()
-    for fr in failed:
-        m = _REWRITE_REPAIR_RE.match(fr.step)
-        if m:
-            iterations.add(int(m.group(1)))
-    return len(iterations)
-
-
-def _resolve_model_hosts(providers: list[ModelProvider]) -> list[str]:
-    """Sorted, deduplicated list of provider host classifications."""
-    return collect_model_hosts([classify_model_host(p) for p in providers])
